IS-1825: Add box with huskelapp info when it already exists
