+++ src/components/huskelapp/Huskelapp.tsx
@@ -1,0 +1,54 @@
+import {
+  HuskelappResponseDTO,
+  oppfolgingsgrunnToText,
+} from "@/data/huskelapp/huskelappTypes";
+import { useRemoveHuskelapp } from "@/data/huskelapp/useRemoveHuskelapp";
+import { BodyShort, Box, Button, Heading, Tooltip } from "@navikt/ds-react";
+import { TrashIcon } from "@navikt/aksel-icons";
+import React from "react";
+
+const texts = {
+  title: "Trenger oppfølging",
+  remove: "Fjern",
+  removeTooltip: "Fjerner huskelappen og oppgaven fra oversikten",
+};
+
+interface ExistingHuskelappProps {
+  huskelapp: HuskelappResponseDTO;
+}
+
+export const Huskelapp = ({ huskelapp }: ExistingHuskelappProps) => {
+  const removeHuskelapp = useRemoveHuskelapp();
+
+  const handleRemoveHuskelapp = (uuid: string) => {
+    removeHuskelapp.mutate(uuid);
+  };
+
+  const existingHuskelappText = !!huskelapp?.tekst
+    ? huskelapp.tekst
+    : !!huskelapp?.oppfolgingsgrunn
+    ? oppfolgingsgrunnToText[huskelapp.oppfolgingsgrunn]
+    : null;
+
+  return (
+    <Box background={"surface-default"} padding="4">
+      <Heading className={"mb-2"} size="small">
+        {texts.title}
+      </Heading>
+      <BodyShort className={"mb-4"}>{existingHuskelappText}</BodyShort>
+      <Tooltip content={texts.removeTooltip}>
+        <Button
+          type="button"
+          icon={<TrashIcon aria-hidden />}
+          variant="danger"
+          onClick={() => handleRemoveHuskelapp(huskelapp.uuid)}
+          loading={removeHuskelapp.isLoading}
+          className={"ml-auto"}
+          size={"small"}
+        >
+          {texts.remove}
+        </Button>
+      </Tooltip>
+    </Box>
+  );
+};
