+++ src/data/huskelapp/huskelappTypes.ts
@@ -1,3 +1,16 @@
+const texts = {
+  oppfolgingsgrunn: {
+    taKontaktSykmeldt: "Ta kontakt med den sykmeldte",
+    taKontaktArbeidsgiver: "Ta kontakt med arbeidsgiver",
+    taKontaktBehandler: "Ta kontakt med behandler",
+    vurderDialogmoteSenere: "Vurder dialogmøte på et senere tidspunkt",
+    folgOppEtterNesteSykmelding: "Følg opp etter neste sykmelding",
+    vurderTiltakBehov: "Vurder behov for tiltak",
+    annet:
+      "Annet (Gi tilbakemelding i Pilotgruppa på Teams hvilken avhukingsvalg du savner)",
+  },
+};
+
 export interface HuskelappRequestDTO {
   oppfolgingsgrunn: Oppfolgingsgrunn;
 }
@@ -18,3 +31,19 @@ export enum Oppfolgingsgrunn {
   VURDER_TILTAK_BEHOV = "VURDER_TILTAK_BEHOV",
   ANNET = "ANNET",
 }
+
+export const oppfolgingsgrunnToText = {
+  [Oppfolgingsgrunn.TA_KONTAKT_SYKEMELDT]:
+    texts.oppfolgingsgrunn.taKontaktSykmeldt,
+  [Oppfolgingsgrunn.TA_KONTAKT_ARBEIDSGIVER]:
+    texts.oppfolgingsgrunn.taKontaktArbeidsgiver,
+  [Oppfolgingsgrunn.TA_KONTAKT_BEHANDLER]:
+    texts.oppfolgingsgrunn.taKontaktBehandler,
+  [Oppfolgingsgrunn.VURDER_DIALOGMOTE_SENERE]:
+    texts.oppfolgingsgrunn.vurderDialogmoteSenere,
+  [Oppfolgingsgrunn.FOLG_OPP_ETTER_NESTE_SYKMELDING]:
+    texts.oppfolgingsgrunn.folgOppEtterNesteSykmelding,
+  [Oppfolgingsgrunn.VURDER_TILTAK_BEHOV]:
+    texts.oppfolgingsgrunn.vurderTiltakBehov,
+  [Oppfolgingsgrunn.ANNET]: texts.oppfolgingsgrunn.annet,
+};
